Backend/requirements.txt
Django==4.2
djangorestframework
PyJWT
psycopg2
pytest
pytest-django
